Complete assignment for JS Scroll
I will have to review the reason for using String.fromCharCode(160)
because I thought  there was an easier way to accomplish this.  Also
where in the assignment description did it instruct us to go to
w3school or other resources to look up the replace() function and the
String.fromCharCode() function?  Javascript has many function so maybe
it would be best to instruct the student to check for existing function
to complete assignment because I was prepared to write my own function
not knowing that replace() function and the String.fromCharCode()
function even existed.